Fix rewiew remarks: test unhealthy, update rst, rename plugin main file
