+++ metrics-scm-1.rockspec
@@ -35,7 +35,7 @@ build = {
         ['metrics.plugins.graphite']                     = 'metrics/plugins/graphite/init.lua',
         ['metrics.plugins.prometheus']                   = 'metrics/plugins/prometheus/init.lua',
         ['metrics.plugins.json']                         = 'metrics/plugins/json/init.lua',
-        ['metrics.plugins.health']                       = 'metrics/plugins/health/init.lua',
+        ['metrics.plugins.health']                       = 'metrics/plugins/health/health.lua',
         ['metrics.default_metrics.tarantool']            = 'metrics/default_metrics/tarantool/init.lua',
         ['metrics.default_metrics.tarantool.fibers']     = 'metrics/default_metrics/tarantool/fibers.lua',
         ['metrics.default_metrics.tarantool.info']       = 'metrics/default_metrics/tarantool/info.lua',
